implemented frequency_modulate() function Tested by inputting its output into frequency_demodulate()
